refactor(snapshots): to their own crate
First pass, still WIP
