+++ cli/Cargo.toml
@@ -19,12 +19,6 @@ name = "deno_bench"
 harness = false
 path = "./bench/main.rs"
 
-[build-dependencies]
-deno_core = { path = "../core", version = "0.86.0" }
-deno_runtime = { path = "../runtime", version = "0.13.0" }
-regex = "1.4.3"
-serde = { version = "1.0.125", features = ["derive"] }
-
 [target.'cfg(windows)'.build-dependencies]
 winapi = "0.3.9"
 winres = "0.1.11"
