+++ runtime/Cargo.toml
@@ -17,23 +17,6 @@ path = "lib.rs"
 name = "hello_runtime"
 path = "examples/hello_runtime.rs"
 
-[build-dependencies]
-deno_console = { path = "../extensions/console", version = "0.5.0" }
-deno_core = { path = "../core", version = "0.86.0" }
-deno_crypto = { path = "../extensions/crypto", version = "0.19.0" }
-deno_fetch = { path = "../extensions/fetch", version = "0.27.0" }
-deno_file = { path = "../extensions/file", version = "0.4.0" }
-deno_timers = { path = "../extensions/timers", version = "0.3.0" }
-deno_url = { path = "../extensions/url", version = "0.5.0" }
-deno_web = { path = "../extensions/web", version = "0.35.0" }
-deno_webgpu = { path = "../extensions/webgpu", version = "0.6.0" }
-deno_webidl = { path = "../extensions/webidl", version = "0.5.0" }
-deno_websocket = { path = "../extensions/websocket", version = "0.10.0" }
-
-[target.'cfg(windows)'.build-dependencies]
-winres = "0.1.11"
-winapi = "0.3.9"
-
 [dependencies]
 deno_console = { path = "../extensions/console", version = "0.5.0" }
 deno_core = { path = "../core", version = "0.86.0" }
