+++ runtime/js.rs
@@ -1,36 +1,36 @@
 // Copyright 2018-2021 the Deno authors. All rights reserved. MIT license.
-use deno_core::Snapshot;
-use log::debug;
+// use deno_core::Snapshot;
+// use log::debug;
 
-pub static CLI_SNAPSHOT: &[u8] =
-  include_bytes!(concat!(env!("OUT_DIR"), "/CLI_SNAPSHOT.bin"));
+// pub static CLI_SNAPSHOT: &[u8] =
+//   include_bytes!(concat!(env!("OUT_DIR"), "/CLI_SNAPSHOT.bin"));
 
-pub fn deno_isolate_init() -> Snapshot {
-  debug!("Deno isolate init with snapshots.");
-  let data = CLI_SNAPSHOT;
-  Snapshot::Static(data)
-}
+// pub fn deno_isolate_init() -> Snapshot {
+//   debug!("Deno isolate init with snapshots.");
+//   let data = CLI_SNAPSHOT;
+//   Snapshot::Static(data)
+// }
 
-#[cfg(test)]
-mod tests {
-  use super::*;
+// #[cfg(test)]
+// mod tests {
+//   use super::*;
 
-  #[test]
-  fn cli_snapshot() {
-    let mut js_runtime = deno_core::JsRuntime::new(deno_core::RuntimeOptions {
-      startup_snapshot: Some(deno_isolate_init()),
-      ..Default::default()
-    });
-    js_runtime
-      .execute(
-        "<anon>",
-        r#"
-      if (!(bootstrap.mainRuntime && bootstrap.workerRuntime)) {
-        throw Error("bad");
-      }
-      console.log("we have console.log!!!");
-    "#,
-      )
-      .unwrap();
-  }
-}
+//   #[test]
+//   fn cli_snapshot() {
+//     let mut js_runtime = deno_core::JsRuntime::new(deno_core::RuntimeOptions {
+//       startup_snapshot: Some(deno_isolate_init()),
+//       ..Default::default()
+//     });
+//     js_runtime
+//       .execute(
+//         "<anon>",
+//         r#"
+//       if (!(bootstrap.mainRuntime && bootstrap.workerRuntime)) {
+//         throw Error("bad");
+//       }
+//       console.log("we have console.log!!!");
+//     "#,
+//       )
+//       .unwrap();
+//   }
+// }
