+++ snapshots/build.rs
@@ -1,0 +1,7 @@
+mod build_runtime;
+mod build_tsc;
+
+fn main() {
+  build_runtime::main();
+  build_tsc::main();
+}
